Rename DemoRecordContainer
